PulseForge/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "psutil>=7.0.0",
    "pynacl>=1.5.0",
    "yt-dlp>=2025.6.25",
]
